requirements.txt
boto3>=1.26.0
neomodel>=5.0.0
mypy-boto3-ec2>=1.26.0
ipaddress>=1.0.23 
pyyaml